backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-generativeai>=0.4.1
groq>=0.9.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
python-multipart>=0.0.6
aiofiles>=23.2.1
slowapi>=0.1.9
reportlab>=4.1.0
gTTS>=2.5.1
pywebpush>=2.0.0
py-vapid>=1.9.1
